+++ lib/utils/get-module-path-for.js
@@ -2,8 +2,6 @@ const fs = require('fs-extra');
 const path = require('path');
 const walkSync = require('walk-sync');
 
-const { getTelemetry } = require('./get-telemetry');
-
 const ADDON_PATHS = {};
 const APP_PATHS = {};
 
@@ -97,20 +95,6 @@ function getModulePathFor(filePath, addonPaths = ADDON_PATHS, appPaths = APP_PAT
   return result;
 }
 
-/**
- * Get the runtime data for the file being transformed
- *
- * @param {String} filePath Absolute path of the file to read data from
- * @returns {Object} Runtime configuration object
- */
-function getTelemetryFor(filePath) {
-  let modulePath = getModulePathFor(filePath);
-  let data = getTelemetry()[modulePath];
-
-  return data;
-}
-
 module.exports = {
-  getTelemetryFor,
   getModulePathFor,
 };
