+++ lib/utils/get-module-path-for.test.js
@@ -1,4 +1,4 @@
-const { getModulePathFor } = require('./get-telemetry-for');
+const { getModulePathFor } = require('./get-module-path-for');
 
 describe('getModulePathFor', () => {
   const addonPaths = {
