+++ lib/utils/telemetry.js
@@ -1,4 +1,5 @@
 const cache = require('../gather/cache');
+const { getModulePathFor } = require('./get-module-path-for');
 
 const CACHE_KEY = 'telemetry';
 
@@ -19,7 +20,21 @@ function getTelemetry() {
   return telemetry;
 }
 
+/**
+ * Get the runtime data for the file being transformed
+ *
+ * @param {String} filePath Absolute path of the file to read data from
+ * @returns {Object} Runtime configuration object
+ */
+function getTelemetryFor(filePath) {
+  let modulePath = getModulePathFor(filePath);
+  let data = getTelemetry()[modulePath];
+
+  return data;
+}
+
 module.exports = {
   getTelemetry,
   setTelemetry,
+  getTelemetryFor,
 };
